+++ src/idt.c
@@ -6,12 +6,12 @@
 struct IdtDescriptorEntry idtTable[IDT_MAX_ENTRIES];
 
 struct IdtDescriptorEntry encodeInterruptDescriptor(void *offset,
-													short segment_selector,
+													uint16_t segment_selector,
 													enum GateType gate_type,
 													char dpl,
 													char p)
 {
-	unsigned long addr = (unsigned long)offset;
+	uint64_t addr = (uint64_t)offset;
 	struct IdtDescriptorEntry entry = {
 		.lower_offset = addr & 0xffff,
 		.segment_selector = segment_selector,
@@ -32,9 +32,9 @@ struct IdtDescriptorEntry encodeInterruptDescriptor(void *offset,
 
 void printIdtDescriptorEntry(struct IdtDescriptorEntry entry)
 {
-	unsigned long full_offset = (unsigned long)entry.lower_offset |
-		((unsigned long)entry.middle_offset << 16) |
-		((unsigned long)entry.higher_offset << 32);
+	uint64_t full_offset = (uint64_t)entry.lower_offset |
+		((uint64_t)entry.middle_offset << 16) |
+		((uint64_t)entry.higher_offset << 32);
 	printk("============================================\n");
 	printk("               IDT entry                     \n");
 	printk("============================================\n");
@@ -59,11 +59,10 @@ void initIdt()
 	initIsr(&idtTable);
 	struct Idtr idtr = {
 		.limit = IDT_MAX_ENTRIES * sizeof(struct IdtDescriptorEntry) - 1,
-		.base_address = (unsigned long long)idtTable
+		.base_address = (uint64_t)idtTable
 	};
 
 	__asm__ volatile("lidt %0" : : "m"(idtr));
 	__asm__ volatile("sti");
 	printk("Interrupts enabled (STI executed)\n");
 }
-
